requirements.txt
typer[all]
rich
python-dotenv
google-generativeai
duckduckgo-search
chardet
python-magic
requests
aiohttp
websockets
psycopg2-binary
sqlalchemy
alembic
python-dotenv
openai==0.28.1
mistralai
ollama
jedi 